+++ service/controller/resource/servicemonitor/resource.go
@@ -19,11 +19,13 @@ const (
 type Config struct {
 	PrometheusClient promclient.Interface
 	Logger           micrologger.Logger
+	Provider         string
 }
 
 type Resource struct {
 	prometheusClient promclient.Interface
 	logger           micrologger.Logger
+	provider         string
 }
 
 func New(config Config) (*Resource, error) {
@@ -33,10 +35,14 @@ func New(config Config) (*Resource, error) {
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
+	if config.Provider == "" {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Provider must not be empty", config)
+	}
 
 	r := &Resource{
 		prometheusClient: config.PrometheusClient,
 		logger:           config.Logger,
+		provider:         config.Provider,
 	}
 
 	return r, nil
@@ -46,16 +52,22 @@ func (r *Resource) Name() string {
 	return Name
 }
 
-func toServiceMonitors(obj interface{}) ([]*promv1.ServiceMonitor, error) {
+func toServiceMonitors(obj interface{}, provider string) ([]*promv1.ServiceMonitor, error) {
 	cluster, err := key.ToCluster(obj)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	return []*promv1.ServiceMonitor{
-		service.APIServer(cluster),
-		service.NginxIngressController(cluster),
-	}, nil
+	serviceMonitors := []*promv1.ServiceMonitor{
+		service.APIServer(cluster, provider),
+		service.NginxIngressController(cluster, provider),
+	}
+
+	if provider == "aws" || provider == "azure" {
+		serviceMonitors = append(serviceMonitors, service.ClusterAutoscaler(cluster, provider))
+	}
+
+	return serviceMonitors, nil
 }
 
 func hasChanged(current, desired *promv1.ServiceMonitor) bool {