+++ service/controller/resource/servicemonitor/service/apiserver.go
@@ -9,7 +9,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
-func APIServer(cluster metav1.Object) *promv1.ServiceMonitor {
+func APIServer(cluster metav1.Object, provider string) *promv1.ServiceMonitor {
 	return &promv1.ServiceMonitor{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      fmt.Sprintf("kubernetes-apiserver-%s", cluster.GetName()),
@@ -28,9 +28,46 @@ func APIServer(cluster metav1.Object) *promv1.ServiceMonitor {
 				Any: true,
 			},
 			Endpoints: []promv1.Endpoint{
-				promv1.Endpoint{
-					Port:   "https",
-					Scheme: "https",
+				{
+					Port:          "https",
+					Scheme:        "https",
+					ScrapeTimeout: "1m",
+					Interval:      "1m",
+					RelabelConfigs: []*promv1.RelabelConfig{
+						{
+							SourceLabels: []string{"__meta_kubernetes_service_name"},
+							TargetLabel:  "app",
+						},
+						{
+							SourceLabels: []string{"__meta_kubernetes_namespace"},
+							TargetLabel:  "namespace",
+						},
+						{
+							SourceLabels: []string{"__meta_kubernetes_pod_name"},
+							TargetLabel:  "pod_name",
+						},
+						{
+							SourceLabels: []string{"__meta_kubernetes_pod_node_name"},
+							TargetLabel:  "node",
+						},
+						{
+							TargetLabel: "cluster_id",
+							Replacement: cluster.GetName(),
+						},
+						{
+							TargetLabel: "cluster_type",
+							Replacement: "tenant_cluster",
+						},
+						{
+							TargetLabel: "provider",
+							Replacement: provider,
+						},
+						{
+							SourceLabels: []string{"__meta_kubernetes_service_label_giantswarm_io_monitoring"},
+							Regex:        "true",
+							Action:       "drop",
+						},
+					},
 					TLSConfig: &promv1.TLSConfig{
 						CAFile:             fmt.Sprintf("/etc/prometheus/secrets/%s/ca", key.Secret()),
 						CertFile:           fmt.Sprintf("/etc/prometheus/secrets/%s/crt", key.Secret()),