+++ service/service.go
@@ -114,6 +114,7 @@ func New(config Config) (*Service, error) {
 			BaseDomain:       config.Viper.GetString(config.Flag.Service.Prometheus.BaseDomain),
 			CreatePVC:        config.Viper.GetBool(config.Flag.Service.Prometheus.Storage.CreatePVC),
 			StorageSize:      config.Viper.GetString(config.Flag.Service.Prometheus.Storage.Size),
+			Provider:         config.Viper.GetString(config.Flag.Service.Provider.Kind),
 		}
 		clusterapiController, err = clusterapi.NewController(c)
 		if err != nil {
@@ -127,10 +128,10 @@ func New(config Config) (*Service, error) {
 			BaseDomain:       config.Viper.GetString(config.Flag.Service.Prometheus.BaseDomain),
 			CreatePVC:        config.Viper.GetBool(config.Flag.Service.Prometheus.Storage.CreatePVC),
 			StorageSize:      config.Viper.GetString(config.Flag.Service.Prometheus.Storage.Size),
+			Provider:         config.Viper.GetString(config.Flag.Service.Provider.Kind),
 			K8sClient:        k8sClient,
 			Logger:           config.Logger,
 			PrometheusClient: prometheusClient,
-			Provider:         config.Viper.GetString(config.Flag.Service.Provider.Kind),
 		}
 		legacyController, err = legacy.NewController(c)
 		if err != nil {