+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Moved prometheus storage from `emptyDir` to a `persistentVolumeClaim`
 
+### Added
+
+- Scraping of legacy docker, calico-node, cluster-autoscaler, aws-node and cadvisor (#78)
+
 ## [0.3.2] - 2020-07-24
 
 ### Changed